Combine resource and getResource file
